requirements.txt
llama-cpp-python
pyttsx3
tk
Pillow
sqlite3